updated embeds & more
apparently i never finished the embeds, so those have been added to all pages. also fixed the 404 CSS issue where if the page you were trying to access didn't exist in the root directory, it would fail to find the CSS file, so now it's deeplinked in the file straight to it via domain.